# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WIB - Second Brain API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Project {
  id: number;
  name: string;
  clientName?: string;
  type: string;
  status: string;
  priority: string;
  description?: string;
  aiSummary?: string;
  progress?: number;
  createdAt: string;
  updatedAt?: string;
}

export interface Task {
  id: number;
  projectId?: number;
  title: string;
  description?: string;
  priority: string;
  status: string;
  dueDate?: string;
  createdAt: string;
  projectName?: string;
}

export interface Person {
  id: number;
  name: string;
  email?: string;
  phone?: string;
  role: string;
  organization?: string;
  createdAt: string;
}

export interface FileRecord {
  id: number;
  projectId?: number;
  name: string;
  fileType: string;
  url?: string;
  size?: number;
  aiSummary?: string;
  createdAt: string;
}

export interface TimelineEvent {
  id: number;
  projectId?: number;
  type: string;
  title: string;
  description?: string;
  projectName?: string;
  createdAt: string;
}

export type ProjectDetail = Project & {
  tasks: Task[];
  people: Person[];
  files: FileRecord[];
  timeline: TimelineEvent[];
};

export interface CreateProjectInput {
  name: string;
  clientName?: string;
  type: string;
  status?: string;
  priority?: string;
  description?: string;
}

export interface UpdateProjectInput {
  name?: string;
  clientName?: string;
  type?: string;
  status?: string;
  priority?: string;
  description?: string;
  progress?: number;
}

export interface CreatePersonInput {
  name: string;
  email?: string;
  phone?: string;
  role: string;
  organization?: string;
}

export interface CreateTaskInput {
  projectId?: number;
  title: string;
  description?: string;
  priority: string;
  status?: string;
  dueDate?: string;
}

export interface UpdateTaskInput {
  title?: string;
  description?: string;
  priority?: string;
  status?: string;
  dueDate?: string;
}

export interface AISuggestions {
  contentType?: string;
  suggestedProject?: string;
  suggestedPeople?: string[];
  priority?: string;
  tasks?: string[];
  tags?: string[];
  summary?: string;
}

export interface InboxItem {
  id: number;
  type: string;
  title: string;
  content?: string;
  rawTranscription?: string;
  aiSuggestions?: AISuggestions;
  status: string;
  fileUrl?: string;
  createdAt: string;
}

export interface ConfirmInboxInput {
  projectId?: number;
  personId?: number;
  createTasks?: boolean;
}

export interface TodaySummary {
  aiSummary?: string;
  urgentTasks: Task[];
  blockedProjects: number;
  pendingInbox: number;
  recentEvents: TimelineEvent[];
}

export interface ChatMessage {
  role: string;
  content: string;
}

export interface ChatRequest {
  message: string;
  history?: ChatMessage[];
}

export interface ChatResponse {
  message: string;
  suggestedActions?: string[];
}

export interface SearchResults {
  projects: Project[];
  people: Person[];
  tasks: Task[];
  files: FileRecord[];
  aiAnswer?: string;
}

export type ListProjectsParams = {
status?: string;
priority?: string;
};

export type ListPeopleParams = {
role?: string;
search?: string;
};

export type ListTasksParams = {
projectId?: number;
priority?: string;
status?: string;
today?: boolean;
};

export type ListInboxItemsParams = {
status?: ListInboxItemsStatus;
};

export type ListInboxItemsStatus = typeof ListInboxItemsStatus[keyof typeof ListInboxItemsStatus];


export const ListInboxItemsStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  dismissed: 'dismissed',
} as const;

export type ListFilesParams = {
projectId?: number;
fileType?: string;
};

export type ListTimelineParams = {
projectId?: number;
limit?: number;
};

export type SearchParams = {
q: string;
};