# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: WIB - Second Brain API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /people:
    get:
      operationId: listPeople
      tags: [people]
      summary: List all people
      parameters:
        - in: query
          name: role
          schema:
            type: string
        - in: query
          name: search
          schema:
            type: string
      responses:
        "200":
          description: List of people
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Person"
    post:
      operationId: createPerson
      tags: [people]
      summary: Create a person
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePersonInput"
      responses:
        "201":
          description: Created person
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Person"
...
  /inbox:
    get:
      operationId: listInboxItems
      tags: [inbox]
      summary: List inbox items
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, confirmed, dismissed]
      responses:
        "200":
          description: List of inbox items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InboxItem"
  /inbox/{id}/confirm:
    post:
      operationId: confirmInboxItem
      tags: [inbox]
      summary: Confirm an inbox item and link to project/people
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfirmInboxInput"
      responses:
        "200":
          description: Confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InboxItem"
  /inbox/{id}/dismiss:
    post:
      operationId: dismissInboxItem
      tags: [inbox]
      summary: Dismiss an inbox item
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Dismissed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InboxItem"
...
  /chat:
    post:
      operationId: sendChat
      tags: [ai]
      summary: Chat with WIB assistant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatRequest"
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatResponse"
  /search:
    get:
      operationId: search
      tags: [ai]
      summary: Intelligent search across all data
      parameters:
        - in: query
          name: q
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResults"
components:
  schemas:
...
    Project:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        clientName:
          type: string
        type:
          type: string
        status:
          type: string
        priority:
          type: string
        description:
          type: string
        aiSummary:
          type: string
        progress:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, name, type, status, priority, createdAt]
...
    Person:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
        organization:
          type: string
        createdAt:
          type: string
      required: [id, name, role, createdAt]
    CreatePersonInput:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
        organization:
          type: string
      required: [name, role]
    Task:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        title:
          type: string
        description:
          type: string
        priority:
          type: string
        status:
          type: string
        dueDate:
          type: string
        createdAt:
          type: string
        projectName:
          type: string
      required: [id, title, priority, status, createdAt]
...
    InboxItem:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
        title:
          type: string
        content:
          type: string
        rawTranscription:
          type: string
        aiSuggestions:
          $ref: "#/components/schemas/AISuggestions"
        status:
          type: string
        fileUrl:
          type: string
        createdAt:
          type: string
      required: [id, type, title, status, createdAt]
    AISuggestions:
      type: object
      properties:
        contentType:
          type: string
        suggestedProject:
          type: string
        suggestedPeople:
          type: array
          items:
            type: string
        priority:
          type: string
        tasks:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        summary:
          type: string
    ConfirmInboxInput:
      type: object
      properties:
        projectId:
          type: integer
        personId:
          type: integer
        createTasks:
          type: boolean
    FileRecord:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        name:
          type: string
        fileType:
          type: string
        url:
          type: string
        size:
          type: integer
        aiSummary:
          type: string
        createdAt:
          type: string
      required: [id, name, fileType, createdAt]
...
    ChatRequest:
      type: object
      properties:
        message:
          type: string
        history:
          type: array
          items:
            $ref: "#/components/schemas/ChatMessage"
      required: [message]
    ChatMessage:
      type: object
      properties:
        role:
          type: string
        content:
          type: string
      required: [role, content]
    ChatResponse:
      type: object
      properties:
        message:
          type: string
        suggestedActions:
          type: array
          items:
            type: string
      required: [message]
    SearchResults:
      type: object
      properties:
        projects:
          type: array
          items:
            $ref: "#/components/schemas/Project"
        people:
          type: array
          items:
            $ref: "#/components/schemas/Person"
        tasks:
          type: array
          items:
            $ref: "#/components/schemas/Task"
        files:
          type: array
          items:
            $ref: "#/components/schemas/FileRecord"
        aiAnswer:
          type: string
      required: [projects, people, tasks, files]