# lib/api-zod/src/generated/types/aISuggestions.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WIB - Second Brain API
 * OpenAPI spec version: 0.1.0
 */

export interface AISuggestions {
  contentType?: string;
  suggestedProject?: string;
  suggestedPeople?: string[];
  priority?: string;
  tasks?: string[];
  tags?: string[];
  summary?: string;
}