# lib/api-zod/src/generated/types/chatRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WIB - Second Brain API
 * OpenAPI spec version: 0.1.0
 */
import type { ChatMessage } from './chatMessage';

export interface ChatRequest {
  message: string;
  history?: ChatMessage[];
}