# lib/api-zod/src/generated/types/inboxItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WIB - Second Brain API
 * OpenAPI spec version: 0.1.0
 */
import type { AISuggestions } from './aISuggestions';

export interface InboxItem {
  id: number;
  type: string;
  title: string;
  content?: string;
  rawTranscription?: string;
  aiSuggestions?: AISuggestions;
  status: string;
  fileUrl?: string;
  createdAt: string;
}